payment api entegration revision 5

+++ app/api/register/route.js
@@ -7,6 +7,7 @@ const merchant_key = process.env.MERCHANT_KEY;
 const merchant_salt = process.env.MERCHANT_SALT;
 
 export async function POST(req) {
+
   try {
     const request = new NextRequest(req);
     const data = await request.json();
@@ -27,8 +28,8 @@ export async function POST(req) {
     console.log(JSON.stringify(data));
 
     // Verify the hash
-    //if (token !== hash) 
-    //  throw new Error('Invalid hash', token, hash);
+    if (token !== hash) 
+      return new Response('Invalid hash: '+token +" "+ hash);
     
 
     const client = await clientPromise;
@@ -37,39 +38,45 @@ export async function POST(req) {
     const purchaseCollection = db.collection('purchase');
 
     // Check if the order exists in the basket
-    //const order = await basketCollection.findOne({ _id: merchant_oid });
-
-    //if (!order) throw new Error('Order not found');
+    const order = await basketCollection.findOne({ _id: merchant_oid });
+    if (!order) return new Response('Order not found');
 
     if (status === 'success') {
-      // Delete the order from the basket collection
-      await basketCollection.deleteOne({ _id: merchant_oid });
 
-      // Add the order to the purchase collection
-      await purchaseCollection.insertOne({
-        ...order,
-        status,
-        total_amount,
-        payment_date: new Date(),
-      });
+      const result = await purchaseCollection.updateOne(
+        { _id: merchant_oid },
+        {
+          $set: {
+            ...order,
+            status,
+            total_amount,
+            payment_date: new Date()
+          },
+        },
+        { upsert: true }
+      );
 
+      if (result.upsertedCount > 0) {
+        // Delete the order from the basket collection
+        await basketCollection.deleteOne({ _id: merchant_oid });
+      }
       // Respond with "OK" to acknowledge the successful payment
       return new Response('OK');
     } else {
       // If payment failed, keep the order in the basket and log the failure reason
       console.error(`Payment failed: ${failed_reason_code} - ${failed_reason_msg}`);
-      return new Response('OK');
+      return new Response('NOT success');
     }
+
   } catch (error) {
     console.error('Server Error:', error);
-    return new Response('Failed to process payment notification', { status: 500 });
+    return new Response('Failed to process payment notification');
   }
 }
-
 
 export async function GET(req) {
 
   console.log("get")
-  return new Response('OK');
+  return new Response('get OK');
 
 }
